Added CreateNewDocument command in the OpenDocumentWindow. Fixed OnDeleteProfile in ProfileManagementViewModel by preventing the deletion of the CurrentProfile. Updated the ProfileManagementViewModel constructor.

+++ INIManagerProject/src/ViewModel/ProfileManagementViewModel.cs
@@ -10,6 +10,7 @@ using System.Threading.Tasks;
 using System.Windows.Controls;
 using System.Windows.Input;
 using System.IO;
+using System.Windows;
 
 namespace INIManagerProject.ViewModel
 {
@@ -23,14 +24,20 @@ namespace INIManagerProject.ViewModel
         public ICommand NewProfile => _newProfile;
         public ICommand DeleteProfile => _deleteProfile;
 
-        public ProfileManagementViewModel(ProfileManager profileManager)
+        public ProfileManagementViewModel()
         {
-            _profileManager = profileManager;
+            // Initialize fields and properties
+            _profileManager = ((App)Application.Current).IniApplication.DocumentManager.CurrentDocument.ProfileManager;
             ProfileList = new ObservableCollection<String>(_profileManager.ProfileList.Select(p => p.ProfileName));
+            // Initialize commands
             _newProfile = new DelegateCommand(OnNewProfile);
             _deleteProfile = new DelegateCommand(OnDeleteProfile);
         }
 
+        /// <summary>
+        /// Create a new profile. The profile name is given by the user.
+        /// </summary>
+        /// <param name="commandParameter">Null</param>
         private void OnNewProfile(object commandParameter)
         {
             var dialog = new InputDialogue("Insert name of new Profile:");
@@ -58,13 +65,20 @@ namespace INIManagerProject.ViewModel
                 _profileManager.CreateNewProfile(profileName);
                 return;
             }
-        }
-         
-        
+        }        
 
+        /// <summary>
+        /// Delete the selected profile
+        /// </summary>
+        /// <param name="commandParameter">The name of the selected profile</param>
         private void OnDeleteProfile(object commandParameter)
         {
             String selectedProfileName = (String)commandParameter;
+            if (selectedProfileName.Equals(_profileManager.CurrentProfile.ProfileName))
+            {
+                MessageBox.Show("Cannot delete current profile!");
+                return;
+            }
             _profileManager.DeleteProfile(selectedProfileName);
             ProfileList.Remove(selectedProfileName);
         }
